v 1.6 убрал лишние действия из алгоритма обработки вызовов

+++ src/main/java/ru/svetozarov/Lift/ILift.java
@@ -4,8 +4,8 @@ package ru.svetozarov.Lift;
  * Created by Evgenij on 19.06.2017.
  */
 public interface ILift extends Runnable{
-    public void callOuter(int floor, boolean route, float idPassenger);
-    public void callInner(int floor, boolean route, float idPassenger);
+    public void callOuter(int floor, boolean route, int idPassenger);
+    public void callInner(int floor, boolean route, int idPassenger);
     public int getFloor();
     public int getState();
     public boolean getRouteCallLift();
